bot/requirements.txt
aiohttp==3.10.11
python-dotenv==1.0.1
rich==13.9.4
click==8.1.7
requests==2.32.3
Pillow==10.2.0
